+++ packages/taquito-ledger-signer/src/error.ts
@@ -1,10 +1,12 @@
+import { ExternalError } from '@taquito/core';
+
 export { InvalidDerivationPathError } from '@taquito/core';
 
 /**
  *  @category Error
  *  @description Error that indicates an invalid or unparseable ledger response
  */
-export class InvalidLedgerResponseError extends Error {
+export class InvalidLedgerResponseError extends ExternalError {
   public name = 'InvalidLedgerResponseError';
   constructor(public message: string) {
     super(message);
@@ -15,7 +17,7 @@ export class InvalidLedgerResponseError extends Error {
  *  @category Error
  *  @description Error that indicates a failure when trying to retrieve a Public Key from Ledger signer
  */
-export class PublicKeyRetrievalError extends Error {
+export class PublicKeyRetrievalError extends ExternalError {
   public name = 'PublicKeyRetrievalError';
   constructor() {
     super(`Unable to retrieve Public Key from Ledger`);
@@ -26,7 +28,7 @@ export class PublicKeyRetrievalError extends Error {
  *  @category Error
  *  @description Error that indicates a failure when trying to retrieve a Public Key Hash from Ledger signer
  */
-export class PublicKeyHashRetrievalError extends Error {
+export class PublicKeyHashRetrievalError extends ExternalError {
   public name = 'PublicKeyHashRetrievalError';
   constructor() {
     super(`Unable to retrieve Public Key Hash from Ledger`);
